+++ docs/CHANGELOG.md
@@ -43,6 +43,7 @@
     * `Allow High-Resolution Title` (Enabled By Default)
     * `Improved Classic Title Positioning` (Enabled By Default)
     * `Use Updated Title` (Enabled By Default)
+    * `Hide Block Outline When GUI Is Hidden` (Enabled By Default)
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`